condence common code for getting change document id
